+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - Added additional tests to test_nodesel, test_heur, and test_strong_branching
 - Migrated documentation to Readthedocs
 - `attachEventHandlerCallback` method to Model for a more ergonomic way to attach event handlers
+- Added optimalNogil, getSolOrigin, retransform, copyModel, solveFirstInterruptOthers
+- Added SCIPsolve nogil, SCIPretransformSol, SCIPtranslateSubSol, SCIPsolGetOrigin, SCIPcopyOrigProb, SCIPcopyOrigVars, SCIPcopyOrigConss, SCIPhashmapCreate, SCIPhashmapFree
+- Added additional tests to test_multi_threads, test_solution, and test_copy
 ### Fixed
 - Fixed too strict getObjVal, getVal check
 ### Changed